fix: overflow, a frozen drag, and a shard we should not complain about
Code review of the branch. Seven findings, all real, all fixed.

The one that mattered: both of §1's run-length forms say "a negative
integer -N", and both were read by negating the entry. -Int.MIN_VALUE is
Int.MIN_VALUE — still negative — so the guards assuming a positive count
did not hold. `"ticks":[-2147483648,[1,2,3]]` drove the write cursor
below zero and indexed out of the buffer, uncaught, inside composition:
a hostile kind 33331/3330/11333 event in a feed crashed the app. The
same hole in facecolors swallowed a run instead of refusing it, and
abs(Int.MIN_VALUE) let a vertex of -2147483648 past the ±64-unit bound,
where *120 overflowed to exactly 0 and the payload parsed valid with the
coordinate silently moved to the origin.

All three are now bounded before they are negated, which also expresses
the real rule: a run longer than the list it repeats into is invalid
whatever its sign. SnoParserOverflowTest pins each, with payloads a
stranger can publish.

The drag rasterised full-resolution on the composition thread — around
48ms a frame at the 1080px a 360.dp viewer asks for on a 3x screen,
which the KDoc had wrongly called "a few hundred pixels square". Now
capped at 512px and moved to Dispatchers.Default, keeping the previous
frame on screen while the next is drawn, so a heavy object makes the
turn coarse rather than freezing the gesture. This is what the plan
asked for and the first pass skipped.

A sealed kind 3330 rendered "could not be read (rule json)" at its
finder — contradicting §7.6, which SnoShardEvent's own KDoc quotes:
failing to open a bag says nothing about the bag. Both 3330s reachable
on a relay are sealed. isSealed() names the case and the renderer draws
nothing for it.

Also: the harness synthesised a 65-character pubkey behind a dead
fallback, and refdriver's vectors path missed the env guard its siblings
have.

+++ quartz/src/commonMain/kotlin/com/vitorpamplona/quartz/cyberspace/deck0003Sno/SnoAvatarWork.kt
@@ -54,9 +54,12 @@ object SnoAvatarWork {
 
     /** The leading zero bits [payload] must have been mined to, as an avatar. */
     fun required(payload: SnoPayload): Int {
-        var farthestTicks = 0
+        // Widened before negating: the parser bounds positions long before this
+        // runs, but a magnitude that is its own negative would price an avatar
+        // at nothing, and this is the one number that decides what gets drawn.
+        var farthestTicks = 0L
         for (tick in payload.positions) {
-            val magnitude = if (tick < 0) -tick else tick
+            val magnitude = if (tick < 0) -tick.toLong() else tick.toLong()
             if (magnitude > farthestTicks) farthestTicks = magnitude
         }
